@@ -19,37 +19,37 @@
     },
     "private": false,
     "dependencies": {
-        "@angular/animations": "12.2.6",
-        "@angular/cdk": "12.2.6",
-        "@angular/common": "12.2.6",
-        "@angular/compiler": "12.2.6",
-        "@angular/core": "12.2.6",
-        "@angular/forms": "12.2.6",
-        "@angular/material": "12.2.6",
-        "@angular/platform-browser": "12.2.6",
-        "@angular/platform-browser-dynamic": "12.2.6",
-        "@angular/router": "12.2.6",
+        "@angular/animations": "13.1.1",
+        "@angular/cdk": "13.1.1",
+        "@angular/common": "13.1.1",
+        "@angular/compiler": "13.1.1",
+        "@angular/core": "13.1.1",
+        "@angular/forms": "13.1.1",
+        "@angular/material": "13.1.1",
+        "@angular/platform-browser": "13.1.1",
+        "@angular/platform-browser-dynamic": "13.1.1",
+        "@angular/router": "13.1.1",
         "@ngx-translate/core": "13.0.0",
         "@ngx-translate/http-loader": "6.0.0",
         "rxjs": "6.6.7",
-        "tslib": "2.2.0",
+        "tslib": "2.3.1",
         "zone.js": "~0.11.4"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~12.2.6",
-        "@angular-eslint/builder": "^12.5.0",
-        "@angular-eslint/eslint-plugin": "^12.5.0",
-        "@angular-eslint/schematics": "^12.5.0",
-        "@angular/cli": "^12.2.6",
-        "@angular/compiler-cli": "^12.2.6",
+        "@angular-devkit/build-angular": "~13.1.2",
+        "@angular-eslint/builder": "^13.0.1",
+        "@angular-eslint/eslint-plugin": "^13.0.1",
+        "@angular-eslint/schematics": "^13.0.1",
+        "@angular/cli": "^13.1.2",
+        "@angular/compiler-cli": "^13.1.1",
         "@types/jasmine": "~3.7.1",
         "@types/jasminewd2": "~2.0.9",
         "@types/node": "^12.11.1",
-        "@typescript-eslint/eslint-plugin": "^4.31.2",
-        "@typescript-eslint/eslint-plugin-tslint": "^4.31.2",
-        "@typescript-eslint/parser": "^4.31.2",
-        "eslint": "^7.32.0",
-        "eslint-plugin-import": "^2.24.2",
+        "@typescript-eslint/eslint-plugin": "5.3.0",
+        "@typescript-eslint/parser": "5.3.0",
+        "eslint": "^8.2.0",
+        "eslint-plugin-jsdoc": "^32.2.0",
+        "eslint-plugin-import": "2.25.2",
         "eslint-plugin-prefer-arrow": "^1.2.3",
         "gulp": "^4.0.2",
         "jasmine-core": "^3.9.0",
@@ -59,10 +59,9 @@
         "karma-coverage-istanbul-reporter": "^3.0.3",
         "karma-jasmine": "^4.0.1",
         "karma-jasmine-html-reporter": "^1.7.0",
-        "ng-packagr": "^12.0.0",
+        "ng-packagr": "^13.1.2",
         "protractor": "~7.0.0",
         "ts-node": "~8.3.0",
-        "tslint": "~6.1.3",
-        "typescript": "~4.3.5"
+        "typescript": "~4.5.4"
     }
 }
