@@ -11,78 +11,112 @@ https://github.com/typescript-eslint/tslint-to-eslint-config/blob/master/docs/FA
 
 Happy linting! 💖
 */
+// eslint:disable
 module.exports = {
     'root': true,
     'ignorePatterns': 'projects/**/*',
     'env': {
         'browser': true,
         'es6': true,
-        'node': true,
+        'node': true
     },
+    'extends': [
+        'plugin:@typescript-eslint/recommended',
+        'plugin:@typescript-eslint/recommended-requiring-type-checking'
+    ],
     'parser': '@typescript-eslint/parser',
     'parserOptions': {
         'project': 'tsconfig.json',
         'sourceType': 'module',
         'createDefaultProgram': true,
     },
-    'extends': [
-        'plugin:@typescript-eslint/recommended',
-        'plugin:@typescript-eslint/recommended-requiring-type-checking',
-    ],
     'plugins': [
-        'eslint-plugin-import',
         '@angular-eslint/eslint-plugin',
+        'eslint-plugin-import',
+        'eslint-plugin-jsdoc',
         'eslint-plugin-prefer-arrow',
-        '@typescript-eslint',
-        '@typescript-eslint/tslint',
+        '@typescript-eslint'
     ],
     'rules': {
+        '@angular-eslint/component-class-suffix': [
+            'error',
+            {
+                'suffixes': [
+                    'Component',
+                    'View',
+                    'Page',
+                    'Dialog',
+                    'Widget'
+                ]
+            }
+        ],
         '@angular-eslint/component-selector': [
             'error',
             {
                 'type': 'element',
-                'prefix': 'volvox',
-                'style': 'kebab-case',
-            },
+                'prefix': [
+                    'volvox'
+                ],
+                'style': 'kebab-case'
+            }
         ],
         '@angular-eslint/directive-class-suffix': 'error',
         '@angular-eslint/directive-selector': [
             'error',
             {
                 'type': 'attribute',
-                'prefix': 'camelCase',
-            },
+                'prefix': [
+                    'volvox'
+                ],
+                'style': 'camelCase'
+            }
         ],
-        '@angular-eslint/no-host-metadata-property': 'error',
         '@angular-eslint/no-input-rename': 'error',
-        '@angular-eslint/no-inputs-metadata-property': 'error',
-        '@angular-eslint/no-output-on-prefix': 'error',
         '@angular-eslint/no-output-rename': 'error',
-        '@angular-eslint/no-outputs-metadata-property': 'error',
         '@angular-eslint/use-pipe-transform-interface': 'error',
-        '@typescript-eslint/unbound-method': [
+        '@typescript-eslint/adjacent-overload-signatures': 'error',
+        '@typescript-eslint/array-type': [
             'error',
             {
-                'ignoreStatic': true,
-            },
+                'default': 'array'
+            }
         ],
-        '@typescript-eslint/restrict-template-expressions': 'off',
-        '@typescript-eslint/no-inferrable-types': 'off',
-        '@typescript-eslint/no-unsafe-call': 'off',
-        '@typescript-eslint/no-unsafe-return': 'off',
-        '@typescript-eslint/no-unsafe-assignment': 'off',
-        '@typescript-eslint/no-var-requires': 'off',
-        '@typescript-eslint/no-unsafe-member-access': 'off',
-        '@typescript-eslint/consistent-type-definitions': 'error',
+        '@typescript-eslint/ban-types': [
+            'error',
+            {
+                'types': {
+                    'Object': {
+                        'message': 'Avoid using the `Object` type. Did you mean `object`?'
+                    },
+                    'Function': {
+                        'message': 'Avoid using the `Function` type. Prefer a specific function type, like `() => void`.'
+                    },
+                    'Boolean': {
+                        'message': 'Avoid using the `Boolean` type. Did you mean `boolean`?'
+                    },
+                    'Number': {
+                        'message': 'Avoid using the `Number` type. Did you mean `number`?'
+                    },
+                    'String': {
+                        'message': 'Avoid using the `String` type. Did you mean `string`?'
+                    },
+                    'Symbol': {
+                        'message': 'Avoid using the `Symbol` type. Did you mean `symbol`?'
+                    },
+                    '{}': false,
+                }
+            }
+        ],
+        '@typescript-eslint/consistent-type-assertions': 'error',
         '@typescript-eslint/dot-notation': 'off',
         '@typescript-eslint/explicit-member-accessibility': [
             'error',
             {
                 'accessibility': 'explicit',
                 'overrides': {
-                    'constructors': 'off',
-                },
-            },
+                    'constructors': 'off'
+                }
+            }
         ],
         '@typescript-eslint/indent': [
             'error',
@@ -91,21 +125,21 @@ module.exports = {
                 'SwitchCase': 1,
                 'FunctionExpression': {
                     'parameters': 'first',
-                },
-            },
+                }
+            }
         ],
         '@typescript-eslint/member-delimiter-style': [
             'error',
             {
                 'multiline': {
                     'delimiter': 'semi',
-                    'requireLast': true,
+                    'requireLast': true
                 },
                 'singleline': {
                     'delimiter': 'comma',
-                    'requireLast': false,
-                },
-            },
+                    'requireLast': false
+                }
+            }
         ],
         '@typescript-eslint/member-ordering': [
             'error',
@@ -181,68 +215,117 @@ module.exports = {
                 ],
             },
         ],
+        '@typescript-eslint/prefer-regexp-exec': 'warn',
+        '@typescript-eslint/restrict-template-expressions': 'off',
+        '@typescript-eslint/restrict-plus-operands': 'warn',
+        '@typescript-eslint/no-unsafe-return': 'off',
         '@typescript-eslint/naming-convention': 'off',
-        '@typescript-eslint/no-empty-function': 'off',
+        '@typescript-eslint/unbound-method': 'off',
+        '@typescript-eslint/no-floating-promises': 'off',
+        '@typescript-eslint/no-empty-function': [
+            'error',
+            {
+                'allow': [
+                    'constructors',
+                ]
+            }
+        ],
         '@typescript-eslint/no-empty-interface': 'error',
+        '@typescript-eslint/no-unsafe-assignment': 'off',
+        '@typescript-eslint/no-unsafe-member-access': 'off',
+        '@typescript-eslint/no-unsafe-argument': 'off',
+        '@typescript-eslint/no-explicit-any': 'warn',
+        '@typescript-eslint/no-inferrable-types': 'off',
         '@typescript-eslint/no-misused-new': 'error',
         '@typescript-eslint/no-misused-promises': 'off',
-        '@typescript-eslint/no-non-null-assertion': 'error',
+        '@typescript-eslint/no-namespace': 'error',
+        '@typescript-eslint/no-parameter-properties': 'off',
         '@typescript-eslint/no-shadow': [
             'error',
             {
-                'hoist': 'all',
-            },
+                'hoist': 'all'
+            }
         ],
-        '@typescript-eslint/no-unused-expressions': [
-            'error',
-            {
-                'allowTernary': true,
-            },
-        ],
+        '@typescript-eslint/no-for-in-array': 'off',
+        '@typescript-eslint/no-unused-expressions': 'off',
+        '@typescript-eslint/no-use-before-define': 'off',
+        '@typescript-eslint/no-var-requires': 'error',
+        '@typescript-eslint/prefer-for-of': 'off',
         '@typescript-eslint/prefer-function-type': 'error',
+        '@typescript-eslint/prefer-namespace-keyword': 'error',
+        '@typescript-eslint/keyword-spacing': 'error',
+        '@typescript-eslint/object-curly-spacing': ['error', 'always'],
         '@typescript-eslint/quotes': [
             'error',
             'single',
+            {
+                'avoidEscape': true
+            }
         ],
         '@typescript-eslint/semi': [
             'error',
-            'always',
+            'always'
+        ],
+        '@typescript-eslint/triple-slash-reference': [
+            'error',
+            {
+                'path': 'always',
+                'types': 'prefer-import',
+                'lib': 'always'
+            }
         ],
         '@typescript-eslint/type-annotation-spacing': 'error',
         '@typescript-eslint/unified-signatures': 'error',
-        'arrow-body-style': 'error',
-        'brace-style': [
-            'error',
-            '1tbs',
-        ],
-        'comma-dangle': [
-            'error',
-            {
-                'objects': 'always-multiline',
-                'arrays': 'always-multiline',
-                'functions': 'always-multiline',
-                'imports': 'always-multiline',
-            },
-        ],
+        '@typescript-eslint/no-unsafe-call': 'off',
+        'arrow-body-style': 'off',
+        'complexity': 'off',
         'constructor-super': 'error',
         'curly': 'error',
-        'eol-last': 'off',
+        'eol-last': 'error',
         'eqeqeq': [
             'error',
-            'smart',
+            'smart'
         ],
-        'guard-for-in': 'error',
-        'id-blacklist': 'off',
-        'id-match': 'off',
+        'guard-for-in': 'off',
+        'id-blacklist': [
+            'error',
+            'any',
+            'Number',
+            'number',
+            'String',
+            'string',
+            'Boolean',
+            'boolean',
+            'Undefined',
+            'undefined'
+        ],
+        'id-match': 'error',
         'import/no-deprecated': 'warn',
-        'import/order': 'error',
+        'jsdoc/check-alignment': 'off',
+        'jsdoc/check-indentation': 'off',
+        'jsdoc/newline-after-description': 'off',
+        'prefer-template': 'error',
+        'max-classes-per-file': [
+            'error',
+            1
+        ],
+        'camelcase': [
+            'error',
+            {
+                'properties': 'never',
+            }
+        ],
         'max-len': [
             'error',
             {
-                'code': 140,
-            },
+                'ignorePattern': '^import',
+                'code': 192
+            }
         ],
+        'new-parens': 'error',
+        'no-bitwise': 'off',
         'no-caller': 'error',
+        'no-cond-assign': 'error',
         'no-console': [
             'error',
             {
@@ -268,66 +351,48 @@ module.exports = {
                     'timeline',
                     'timelineEnd',
                     'timeStamp',
-                    'context',
-                ],
-            },
+                    'context'
+                ]
+            }
         ],
         'no-debugger': 'error',
-        'no-empty': 'off',
+        'no-empty': 'error',
         'no-eval': 'error',
         'no-fallthrough': 'error',
+        'no-invalid-this': 'off',
         'no-new-wrappers': 'error',
-        'no-restricted-imports': [
-            'error',
-            'rxjs/Rx',
-        ],
         'no-throw-literal': 'error',
         'no-trailing-spaces': 'error',
         'no-undef-init': 'error',
         'no-underscore-dangle': 'off',
+        'no-unsafe-finally': 'error',
         'no-unused-labels': 'error',
         'no-var': 'error',
+        'object-shorthand': 'error',
+        'one-var': [
+            'error',
+            'never'
+        ],
+        'prefer-arrow/prefer-arrow-functions': 'off',
         'prefer-const': 'error',
-        'radix': 'off',
-        'object-curly-spacing': [ 'error', 'always' ],
-        'computed-property-spacing': [ 'error', 'always' ],
+        'radix': 'error',
+        'space-before-function-paren': [
+            'error',
+            {
+                'anonymous': 'never',
+                'asyncArrow': 'always',
+                'named': 'never'
+            }
+        ],
         'spaced-comment': [
             'error',
             'always',
             {
-                'exceptions': [ '*' ],
-                'markers': [
-                    '/', '#region', '#endregion',
-                ],
-            },
+                'exceptions': ['*'],
+                'markers': ['/', '#region', '#endregion'],
+            }
         ],
-        '@typescript-eslint/tslint/config': [
-            'error',
-            {
-                'rules': {
-                    'import-spacing': true,
-                    'typedef': [
-                        true,
-                        'call-signature',
-                        'arrow-call-signature',
-                        'parameter',
-                        'arrow-parameter',
-                        'property-declaration',
-                        'member-variable-declaration',
-                        'object-destructuring',
-                        'array-destructuring',
-                    ],
-                    'use-life-cycle-interface': true,
-                    'whitespace': [
-                        true,
-                        'check-branch',
-                        'check-decl',
-                        'check-operator',
-                        'check-separator',
-                        'check-type',
-                    ],
-                },
-            },
-        ],
-    },
+        'use-isnan': 'error',
+        'valid-typeof': 'off',
+    }
 };
