@@ -23,8 +23,9 @@ export class VolvoxSocketService {
         this.subject.next(data);
     }
 
-    private create(url: string): Subject<MessageEvent> {
+    private create<T>(url: string): Subject<MessageEvent> {
         const ws = new WebSocket(url);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
         const observable = new Observable((obs: Observer<MessageEvent>): any => {
             ws.onmessage = obs.next.bind(obs);
             ws.onerror = obs.error.bind(obs);
@@ -32,6 +33,7 @@ export class VolvoxSocketService {
             return ws.close.bind(ws);
         });
         const observer = {
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
             next: (data: any): void => {
                 if (ws.readyState === WebSocket.OPEN) {
                     ws.send(JSON.stringify(data));
