@@ -1,5 +1,7 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { BaseComponent } from '@volvox-ng/material';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';<% if (createFacade){ %>
+import { Observable } from 'rxjs';
+import { <%= classify(name) %>Facade } from '<%= facadeRelativePath %>';
+import { <% if (!useClass) { %>I<% } %><%= classify(name) %>State } from '<%= facadeModelFolderRelativePath %>';<% } %>
 
 @Component({
     selector: '<%= dasherize(prefix) %>-<%= dasherize(name) %>',
@@ -7,14 +9,15 @@ import { BaseComponent } from '@volvox-ng/material';
     styleUrls: [ './<%= dasherize(name) %>.component.<%= style %>' ],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class <%= classify(name) %>Component extends BaseComponent implements OnInit {
-
-    constructor() {
-        super();
+export class <%= classify(name) %>Component implements OnInit {
+<% if (createFacade && useFacadeInTemplate) { %>
+    public <%= camelize(name) %>State$: Observable<<% if (!useClass) { %>I<% } %><%= classify(name) %>State>;
+<% } %>
+    constructor(<% if (createFacade) { %>private my<%= classify(name) %>Facade: <%= classify(name) %>Facade<% } %>) {
     }
 
-    public ngOnInit(): void {
-        super.ngOnInit();
+    public ngOnInit(): void {<% if (createFacade) {%>
+        this.<%= camelize(name) %>State$ = this.my<%= classify(name) %>Facade.subState();<% } %>
     }
 
 }
