@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { I<%= classify(name) %>State } from './<%= dasherize(name) %>-state.model';
+import { <% if (!useClass) { %>I<% } %><%= classify(name) %>State } from '<%= modelRelativePath %>';
 
-let _state: I<%= classify(name) %>State = {
+let _state: <% if (!useClass) { %>I<% } %><%= classify(name) %>State = {
 };
 
 @Injectable({
@@ -10,21 +10,23 @@ let _state: I<%= classify(name) %>State = {
 })
 export class <%= classify(name) %>Facade {
 
-    protected store$: BehaviorSubject<I<%= classify(name) %>State>;
+    protected store$: BehaviorSubject<<% if (!useClass) { %>I<% } %><%= classify(name) %>State> = new BehaviorSubject<<% if (!useClass) { %>I<% } %><%= classify(name) %>State>(_state);
 
     constructor() {
-        this.store$ = new BehaviorSubject<I<%= classify(name) %>State>(_state);
     }
 
-    public subState(): Observable<I<%= classify(name) %>State> {
+    public subState(): Observable<<% if (!useClass) { %>I<% } %><%= classify(name) %>State> {
         return this.store$.asObservable();
     }
-
-    public get snapshot(): I<%= classify(name) %>State {
+<% if (withSnapshot) { %>
+    <% if (useSnapshotFunction) { %>public getStateSnapshot(): <% if (!useClass) { %>I<% } %><%= classify(name) %>State {
         return this.store$.value;
-    }
-
-    private updateState(state: I<%= classify(name) %>State): void {
+    }<% } else { %>public get snapshot(): <% if (!useClass) { %>I<% } %><%= classify(name) %>State {
+        return this.store$.value;
+    }<% } %>
+<% } %>
+    private updateState(state: <% if (!useClass) { %>I<% } %><%= classify(name) %>State): void {
         this.store$.next(_state = state);
     }
+
 }
