@@ -1,7 +1,7 @@
-import { SourceFile } from 'typescript';
+import * as ts from "@schematics/angular/third_party/github.com/Microsoft/TypeScript/lib/typescript";
 
 export interface IModuleContext {
-    source: SourceFile;
+    source: ts.SourceFile;
     relativePath: string;
     classifiedName: string;
 }
