@@ -1,1 +1,1 @@
-export type StyleSchema = 'css' |'scss' | 'sass' | 'less';
+export type StyleSchema = 'css' | 'scss' | 'sass' | 'less';
