@@ -2,7 +2,7 @@ import { Rule, SchematicContext, Tree } from '@angular-devkit/schematics';
 import { NodePackageInstallTask } from '@angular-devkit/schematics/tasks';
 
 export function ngAdd(): Rule {
-    return (tree: Tree, context: SchematicContext): Tree => {
+    return (tree: Tree, context: SchematicContext) => {
         context.addTask(new NodePackageInstallTask());
         return tree;
     };
